requirements.txt
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pytest>=7.3.1
jupyter>=1.0.0
pyyaml>=6.0
gitpython>=3.1.30
beautifulsoup4>=4.12.0
requests>=2.28.2
networkx>=3.0
tqdm>=4.65.0 